requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
asyncpg==0.30.0
psycopg[binary]==3.2.5
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose==3.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
redis==5.2.1
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
